requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
psycopg2-binary==2.9.9
sentence-transformers==3.1.1
google-generativeai==0.8.3
python-dotenv==1.0.1
langgraph==0.2.14
langchain==0.3.1
langchain-community==0.3.0